feat: create hook usePlanet and use it in plantInfo component
